Reflect Review

+++ Lib9c/TableData/Cost/EnhancementCostSheetV2.cs
@@ -11,38 +11,38 @@ namespace Nekoyume.TableData
     {
         public class Row : EnhancementCostSheet.Row
         {
-            public decimal SuccessRatio { get; private set; }
-            public decimal GreatSuccessRatio { get; private set; }
-            public decimal FailRatio { get; private set; }
+            public int SuccessRatio { get; private set; }
+            public int GreatSuccessRatio { get; private set; }
+            public int FailRatio { get; private set; }
             public int SuccessRequiredBlockIndex { get; private set; }
             public int GreatSuccessRequiredBlockIndex { get; private set; }
             public int FailRequiredBlockIndex { get; private set; }
-            public decimal BaseStatGrowthMin { get; private set; }
-            public decimal BaseStatGrowthMax { get; private set; }
-            public decimal ExtraStatGrowthMin { get; private set; }
-            public decimal ExtraStatGrowthMax { get; private set; }
-            public decimal ExtraSkillDamageGrowthMin { get; private set; }
-            public decimal ExtraSkillDamageGrowthMax { get; private set; }
-            public decimal ExtraSkillChanceGrowthMin { get; private set; }
-            public decimal ExtraSkillChanceGrowthMax { get; private set; }
+            public int BaseStatGrowthMin { get; private set; }
+            public int BaseStatGrowthMax { get; private set; }
+            public int ExtraStatGrowthMin { get; private set; }
+            public int ExtraStatGrowthMax { get; private set; }
+            public int ExtraSkillDamageGrowthMin { get; private set; }
+            public int ExtraSkillDamageGrowthMax { get; private set; }
+            public int ExtraSkillChanceGrowthMin { get; private set; }
+            public int ExtraSkillChanceGrowthMax { get; private set; }
 
             public override void Set(IReadOnlyList<string> fields)
             {
                 base.Set(fields);
-                SuccessRatio = ParseDecimal(fields[5], 0m);
-                GreatSuccessRatio = ParseDecimal(fields[6], 0m);
-                FailRatio = ParseDecimal(fields[7], 0m);
+                SuccessRatio = ParseInt(fields[5], 0);
+                GreatSuccessRatio = ParseInt(fields[6], 0);
+                FailRatio = ParseInt(fields[7], 0);
                 SuccessRequiredBlockIndex = ParseInt(fields[8], 0);
                 GreatSuccessRequiredBlockIndex = ParseInt(fields[9], 0);
                 FailRequiredBlockIndex = ParseInt(fields[10], 0);
-                BaseStatGrowthMin = ParseDecimal(fields[11], 0m);
-                BaseStatGrowthMax = ParseDecimal(fields[12], 0m);
-                ExtraStatGrowthMin = ParseDecimal(fields[13], 0m);
-                ExtraStatGrowthMax = ParseDecimal(fields[14], 0m);
-                ExtraSkillDamageGrowthMin = ParseDecimal(fields[15], 0m);
-                ExtraSkillDamageGrowthMax = ParseDecimal(fields[16], 0m);
-                ExtraSkillChanceGrowthMin = ParseDecimal(fields[17], 0m);
-                ExtraSkillChanceGrowthMax = ParseDecimal(fields[18], 0m);
+                BaseStatGrowthMin = ParseInt(fields[11], 0);
+                BaseStatGrowthMax = ParseInt(fields[12], 0);
+                ExtraStatGrowthMin = ParseInt(fields[13], 0);
+                ExtraStatGrowthMax = ParseInt(fields[14], 0);
+                ExtraSkillDamageGrowthMin = ParseInt(fields[15], 0);
+                ExtraSkillDamageGrowthMax = ParseInt(fields[16], 0);
+                ExtraSkillChanceGrowthMin = ParseInt(fields[17], 0);
+                ExtraSkillChanceGrowthMax = ParseInt(fields[18], 0);
             }
         }
